test(binding-kafka): move KafkaPipelineBM into a bench subpackage
KafkaController, KafkaSource, KafkaSink, and KafkaEvent are now public.
internal.cache is not exported by module-info.java, so this doesn't
widen the module's real external surface, but it lets benchmark classes
live in a normal bench/ subpackage instead of alongside the production
sources they exercise.

KafkaPipelineBM moves into internal.cache.bench accordingly; its B/op
and throughput figures are unchanged by the move.
